requirements.txt
python-dotenv==1.0.1
langgraph==0.2.40
openai==1.52.2
faster-whisper==1.0.3
scenedetect[opencv]==0.6.2
opencv-python==4.10.0.84
imageio-ffmpeg==0.5.1
httpx==0.27